chi-explorer/requirements.txt
Flask==3.0.3
requests==2.32.3
matplotlib==3.9.2
numpy==2.0.2
json-repair==0.30.3
python-dotenv==1.0.1
